updated npm packages and resolved problems from that
