+++ examples/us_enrichment.mjs
@@ -1,6 +1,7 @@
-const SmartySDK = require("smartystreets-javascript-sdk");
+import SmartySDK from "smartystreets-javascript-sdk";
+
 const SmartyCore = SmartySDK.core;
-const Lookup = SmartySDK.usReverseGeo.Lookup;
+const Lookup = SmartySDK.usEnrichment.Lookup;
 
 // for Server-to-server requests, use this code:
 // let authId = process.env.SMARTY_AUTH_ID;
@@ -14,29 +15,25 @@ const credentials = new SmartyCore.SharedCredentials(key);
 // The appropriate license values to be used for your subscriptions
 // can be found on the Subscription page of the account dashboard.
 // https://www.smarty.com/docs/cloud/licensing
-let clientBuilder = new SmartyCore.ClientBuilder(credentials).withLicenses(["us-reverse-geocoding-cloud"]);
+let clientBuilder = new SmartyCore.ClientBuilder(credentials).withLicenses(["us-property-data-principal-cloud"]);
 	// .withBaseUrl("YOUR URL") // withBaseUrl() should be used if you are self-hosting the Smarty API
-let client = clientBuilder.buildUsReverseGeoClient();
 
-let lookup1 = new Lookup(40.27644, -111.65747);
+let client = clientBuilder.buildUsEnrichmentClient();
+
+// Documentation for input fields can be found at:
+// https://www.smarty.com/docs/us-street-api#input-fields
+
+let lookup = new Lookup("334968275");
 // uncomment the following line to add a custom parameter
-// lookup1.addCustomParameter("source", "all");
+// lookup.addCustomParameter("include", "group_financial");
 
-await handleResponse(lookup1);
-
-function displayResult(result) {
-	console.log(result.response.results[0].address);
-}
-
-function handleError(error) {
-	console.log("ERROR:", error);
-}
+handleResponse(lookup).then();
 
 async function handleResponse(lookup) {
 	try {
-		const result = await client.send(lookup);
-		displayResult(result);
-	} catch(err) {
-		handleError(err);
+		const result = await client.sendPrincipal(lookup);
+		console.log(result.response);
+	} catch (err) {
+		console.log(err);
 	}
 }
