+++ package.json
@@ -56,7 +56,7 @@
   "@rollup/plugin-typescript": "^12.3.0",
   "@types/chai": "^5.2.3",
   "@types/mocha": "^10.0.10",
-  "@types/node": "^24.9.1",
+  "@types/node": "^24.9.2",
   "chai": "^6.2.0",
   "mocha": "^11.7.4",
   "prettier": "3.6.2",
@@ -66,8 +66,9 @@
   "typescript": "^5.9.3"
  },
  "dependencies": {
-  "axios": "^1.13.0",
+  "axios": "^1.13.1",
   "axios-retry": "^4.5.0",
+  "smartystreets-javascript-sdk": "^7.1.0",
   "tslib": "^2.8.1"
  }
 }
